+++ modules/alloy/src/main/ts/ephox/alloy/positioning/mode/SelectionAnchor.ts
@@ -1,5 +1,5 @@
 import { FieldSchema } from '@ephox/boulder';
-import { Optional, Unicode } from '@ephox/katamari';
+import { Arr, Optional, Unicode } from '@ephox/katamari';
 import { Insert, Remove, SimRange, SimSelection, SugarElement, SugarNode, Traverse, WindowSelection } from '@ephox/sugar';
 
 import * as Descend from '../../alien/Descend';
@@ -15,14 +15,21 @@ import * as ContentAnchorCommon from './ContentAnchorCommon';
 const descendOnce = (element: SugarElement<Node>, offset: number): Descend.ElementAndOffset<Node> =>
   SugarNode.isText(element) ? Descend.point(element, offset) : Descend.descendOnce(element, offset);
 
-const getAnchorSelection = (win: Window, anchorInfo: SelectionAnchor): Optional<SimRange> => {
+const isSimRange = (detail: SimRange | SugarElement<Node>[]): detail is SimRange =>
+  (detail as SimRange).start !== undefined;
+
+const getAnchorSelection = (win: Window, anchorInfo: SelectionAnchor): Optional<SimRange | SugarElement<HTMLTableCellElement>[]> => {
   // FIX TEST Test both providing a getSelection and not providing a getSelection
   const getSelection = anchorInfo.getSelection.getOrThunk(() => () => WindowSelection.getExact(win));
 
   return getSelection().map((sel) => {
-    const modStart = descendOnce(sel.start, sel.soffset);
-    const modFinish = descendOnce(sel.finish, sel.foffset);
-    return SimSelection.range(modStart.element, modStart.offset, modFinish.element, modFinish.offset);
+    if (isSimRange(sel)) {
+      const modStart = descendOnce(sel.start, sel.soffset);
+      const modFinish = descendOnce(sel.finish, sel.foffset);
+      return SimSelection.range(modStart.element, modStart.offset, modFinish.element, modFinish.offset);
+    }
+
+    return sel;
   });
 };
 
@@ -32,20 +39,49 @@ const placement = (component: AlloyComponent, anchorInfo: SelectionAnchor, origi
 
   const selectionBox = getAnchorSelection(win, anchorInfo).bind((sel) => {
     // This represents the *visual* rectangle of the selection.
-    const optRect = WindowSelection.getBounds(win, SimSelection.exactFromRange(sel)).orThunk(() => {
-      const x = SugarElement.fromText(Unicode.zeroWidth);
-      Insert.before(sel.start, x);
-      // Certain things like <p><br/></p> with (p, 0) or <br>) as collapsed selection do not return a client rectangle
-      const rect = WindowSelection.getFirstRect(win, SimSelection.exact(x, 0, x, 1));
-      Remove.remove(x);
-      return rect;
-    });
+    if (isSimRange(sel)) {
+      const optRect = WindowSelection.getBounds(win, SimSelection.exactFromRange(sel)).orThunk(() => {
+        const x = SugarElement.fromText(Unicode.zeroWidth);
+        Insert.before(sel.start, x);
+        // Certain things like <p><br/></p> with (p, 0) or <br>) as collapsed selection do not return a client rectangle
+        const rect = WindowSelection.getFirstRect(win, SimSelection.exact(x, 0, x, 1));
+        Remove.remove(x);
+        return rect;
+      });
 
-    return optRect.bind((rawRect) => ContentAnchorCommon.getBox(rawRect.left, rawRect.top, rawRect.width, rawRect.height));
+      return optRect.bind((rawRect) => {
+        return ContentAnchorCommon.getBox(rawRect.left, rawRect.top, rawRect.width, rawRect.height);
+      });
+    } else {
+      const menuRects = sel.map((x) => x.dom.getBoundingClientRect());
+
+      const minMax = Arr.foldl(menuRects, (acc, rect) => {
+        return {
+          left: Math.min(acc.left, rect.left),
+          right: Math.max(acc.right, rect.right),
+          top: Math.min(acc.top, rect.top),
+          bottom: Math.max(acc.bottom, rect.bottom)
+        };
+      },
+      {
+        left: Number.MAX_VALUE,
+        right: Number.MIN_VALUE,
+        top: Number.MAX_VALUE,
+        bottom: Number.MIN_VALUE
+      });
+
+      return ContentAnchorCommon.getBox(minMax.left, minMax.top, minMax.right - minMax.left, minMax.bottom - minMax.top);
+    }
   });
 
   const targetElement = getAnchorSelection(win, anchorInfo)
-    .bind((sel) => SugarNode.isElement(sel.start) ? Optional.some(sel.start) : Traverse.parentElement(sel.start));
+    .bind((sel) => {
+      if (isSimRange(sel)) {
+        return SugarNode.isElement(sel.start) ? Optional.some(sel.start) : Traverse.parentElement(sel.start);
+      }
+
+      return Arr.head(sel);
+    });
   const elem = targetElement.getOr(component.element);
 
   return ContentAnchorCommon.calcNewAnchor(selectionBox, rootPoint, anchorInfo, origin, elem);