+++ modules/tinymce/src/plugins/quickbars/test/ts/module/test/Utils.ts
@@ -1,5 +1,5 @@
 import { UiFinder, Waiter } from '@ephox/agar';
-import { SugarBody } from '@ephox/sugar';
+import { SugarBody, SugarElement } from '@ephox/sugar';
 
 const toolbarSelector = '.tox-pop__dialog .tox-toolbar';
 
@@ -9,7 +9,14 @@ const pAssertToolbarNotVisible = (): Promise<void> =>
 const pAssertToolbarVisible = (): Promise<void> =>
   Waiter.pTryUntil('toolbar should exist', () => UiFinder.exists(SugarBody.body(), toolbarSelector));
 
+const pGetToolBar = async (): Promise<SugarElement<Element>> => {
+  return Waiter.pTryUntil('get toolbar',
+    () => UiFinder.findIn(SugarBody.body(), toolbarSelector).getOrDie()
+  );
+};
+
 export {
   pAssertToolbarNotVisible,
-  pAssertToolbarVisible
+  pAssertToolbarVisible,
+  pGetToolBar
 };