+++ modules/tinymce/src/themes/silver/main/ts/backstage/Anchors.ts
@@ -1,5 +1,5 @@
 import { AlloyComponent, Bubble, HotspotAnchorSpec, Layout, LayoutInset, MaxHeight, NodeAnchorSpec, SelectionAnchorSpec } from '@ephox/alloy';
-import { Optional } from '@ephox/katamari';
+import { Arr, Optional } from '@ephox/katamari';
 import { SimSelection, SugarElement, SugarShadowDom } from '@ephox/sugar';
 
 import Editor from 'tinymce/core/api/Editor';
@@ -84,6 +84,13 @@ const getCursorAnchor = (editor: Editor, bodyElement: () => SugarElement<HTMLEle
   root: bodyElement(),
   getSelection: () => {
     const rng = editor.selection.getRng();
+
+    // Only return a range if there is a selection of more than one cell.
+    const selectedCells = editor.model.table.getSelectedCells();
+    if (selectedCells.length > 1) {
+      return Optional.some(Arr.map(selectedCells, SugarElement.fromDom));
+    }
+
     return Optional.some(
       SimSelection.range(SugarElement.fromDom(rng.startContainer), rng.startOffset, SugarElement.fromDom(rng.endContainer), rng.endOffset)
     );