+++ modules/alloy/CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - Select would in some situations when it was reused or otherwise reset not have an initial selected value when expected to. #TINY-9679
 
+### Improved
+- `Anchor` now returns the selected cells, if the selection contains more than one table cell. #TINY-8297
+- `SelectionAnchor` now uses the bounds of table cell selection, if the selection contains more than one table cell. #TINY-8297
+
 ## 12.1.0 - 2023-03-15
 
 ### Added